♻️ refactor(system): move system page into feature module
Move the system information page implementation and styles under src/features/system.

Keep src/pages/SystemPage.tsx as a route wrapper to maintain the frontend architecture boundary.
